+++ pepper-apis/src/main/java/org/broadinstitute/ddp/db/dao/JdbiWorkflowState.java
@@ -3,7 +3,9 @@ package org.broadinstitute.ddp.db.dao;
 import java.util.Optional;
 
 import org.broadinstitute.ddp.model.workflow.StateType;
+import org.broadinstitute.ddp.model.workflow.StudyRedirectState;
 import org.jdbi.v3.sqlobject.SqlObject;
+import org.jdbi.v3.sqlobject.config.RegisterConstructorMapper;
 import org.jdbi.v3.sqlobject.customizer.Bind;
 import org.jdbi.v3.sqlobject.statement.GetGeneratedKeys;
 import org.jdbi.v3.sqlobject.statement.SqlQuery;
@@ -16,8 +18,19 @@ public interface JdbiWorkflowState extends SqlObject {
     @GetGeneratedKeys
     long insert(@Bind("type") StateType type);
 
+    @SqlUpdate("insert into workflow_study_redirect_state (workflow_state_id, study_guid, redirect_url) "
+            + " values (:stateId, :studyGuid, :redirectUrl)")
+    int insert(@Bind("stateId") long stateId, @Bind("studyGuid") String studyGuid, @Bind("redirectUrl") String redirectUrl);
+
     @SqlQuery("select ws.workflow_state_id from workflow_state as ws"
             + " join workflow_state_type as wst on wst.workflow_state_type_id = ws.workflow_state_type_id"
             + " where wst.workflow_state_type_code = :type")
     Optional<Long> findIdByType(@Bind("type") StateType type);
+
+    @SqlQuery("select workflow_state_id, study_guid, redirect_url from workflow_study_redirect_state "
+            + " where study_guid = :studyGuid and redirect_url = :redirectUrl")
+    @RegisterConstructorMapper(StudyRedirectState.class)
+    Optional<StudyRedirectState> findIdByStudyNameGuidAndRedirectUrl(@Bind("studyGuid") String studyGuid,
+                                                                     @Bind("redirectUrl") String redirectUrl);
+
 }
