+++ pepper-apis/src/main/java/org/broadinstitute/ddp/json/workflow/WorkflowResponse.java
@@ -8,16 +8,6 @@ public class WorkflowResponse {
     @SerializedName("next")
     protected String next;
 
-    @SerializedName("studyName")
-    protected String studyName;
-
-    @SerializedName("studyGuid")
-    protected String studyGuid;
-
-    @SerializedName("redirectUrl")
-    protected String redirectUrl;
-
-
     public static WorkflowResponse from(StaticState state) {
         WorkflowResponse resp = new WorkflowResponse();
         resp.next = state.getType().name();
@@ -30,30 +20,11 @@ public class WorkflowResponse {
         return resp;
     }
 
-    public WorkflowResponse(String next, String studyName, String studyGuid, String redirectUrl) {
-        this.next = next;
-        this.studyName = studyName;
-        this.studyGuid = studyGuid;
-        this.redirectUrl = redirectUrl;
-    }
-
     protected WorkflowResponse() {
         // Use static factories.
     }
 
     public String getNext() {
         return next;
-    }
-
-    public String getStudyName() {
-        return studyName;
-    }
-
-    public String getStudyGuid() {
-        return studyGuid;
-    }
-
-    public String getRedirectUrl() {
-        return redirectUrl;
     }
 }
