+++ study-builder/studies/pancan/i18n/en.conf
@@ -3,33 +3,33 @@
 "prequal_translatedTitles"="Lets get started"
 "prequal_translatedSubTitles"="Thank you for your interest in Count Me In. <br> Here's what sign up and participation looks like:"
 "prequal_intro"="Thank you for your interest in Count Me In. <br> Here's what sign up and participation looks like:"
-"prequal_introduction_blocks_01_bodyTemplate_01"="Registration - Answer a few short questions about yourself or your child (3m)"
-"prequal_introduction_blocks_01_bodyTemplate_02"="Consent Form - Read through and sign the consent form (10-15m)"
-"prequal_introduction_blocks_01_bodyTemplate_03"="Medical Release Form - Tell us where you or your child have been treated (5m)"
-"prequal_introduction_blocks_01_bodyTemplate_04"="Surveys - Answer questions about you or your child's experience with cancer (10-15m)"
+"prequal_introduction_registration"="Registration - Answer a few short questions about yourself or your child (3m)"
+"prequal_introduction_consentForm"="Consent Form - Read through and sign the consent form (10-15m)"
+"prequal_introduction_medicalReleaseForm"="Medical Release Form - Tell us where you or your child have been treated (5m)"
+"prequal_introduction_surveyAnswers"="Surveys - Answer questions about you or your child's experience with cancer (10-15m)"
 
-"prequal_introduction_blocks_01_bodyTemplate_05"="Your progress will be automatically saved so you can come back to any of these steps later."
+"prequal_introduction_progress"="Your progress will be automatically saved so you can come back to any of these steps later."
 
-"prequal_sections_01_blocks_01_title_01"="First who is signing up for Count Me In"
-"prequal_sections_01_blocks_01_subtitle_01"="<i>Check all that apply</i>"
+"prequal_introduction_titleWho"="First who is signing up for Count Me In"
+"prequal_Check_that_apply"="<i>Check all that apply</i>"
 
-"prequal_sections_01_question_01_picklistOptions_01_optionLabelTemplate"="I have been diagnosed with cancer and signing up for myself"
-"prequal_sections_01_question_01_picklistOptions_02_optionLabelTemplate"="My child has been diagnosed with cancer and signing up with them or for them"
-"prequal_sections_01_question_01_validations_hintTemplate"="Please choose one of the above options."
+"prequal_selfDiagnosed"="I have been diagnosed with cancer and signing up for myself"
+"prequal_childDiagnosed"="My child has been diagnosed with cancer and signing up with them or for them"
+"prequal_choose_option"="Please choose one of the above options."
 
-"prequal_diagnosed_primary_cancer_list_prompt"="What primary cancer(s) have you/your child been diagnosed with? \n If you/your child have been diagnosed with more than one cancer type, please use the button below to add an additional diagnosis."
-"prequal_sections_02_question_01_validations_01_hintTemplate_01"="If you/your child have been diagnosed with more than one cancer type, please use the button below to add an additional diagnosis."
+"prequal_diagnosedPrimaryCancerList"="What primary cancer(s) have you/your child been diagnosed with? <br/> If you/your child have been diagnosed with more than one cancer type, please use the button below to add an additional diagnosis."
+"prequal_diagnosedMultipleCancerList"="If you/your child have been diagnosed with more than one cancer type, please use the button below to add an additional diagnosis."
 
-"prequal_diagnosed_primary_cancer_list_question_additionalItemTemplate":""
-"prequal_diagnosed_primary_cancer_list_add_buttonTemplate": "+ add another diagnosis"
+"prequal_diagnose_cancer_list_additionalTmpl":""
+"prequal_diagnosed_cancer_list_additional_diagnosis": "+ add another diagnosis"
 "prequal_diagnosed_primary_cancer_advanced":"Have you been diagnosed with metastatic breast cancer (also known as stage IV breast cancer)?",
-"prequal_diagnosed_primary_cancer_prompt": "Select Diagnosis"
-"prequal_diagnosed_primary_cancer_breast_advanced":"Have you been diagnosed with metastatic breast cancer (also known as stage IV breast cancer)?",
-"prequal_diagnosed_primary_cancer_prostate_advanced":"Have you been diagnosed with metastatic prostate cancer (also known as prostate cancer that has spread beyond the prostate, including biochemical recurrence)?",
-"prequal_self_age_prompt":"How old are you?"
-"prequal_child_age_prompt":"How old is your child?"
-"prequal_age_placeholder":"Enter Age"
-"prequal_age_hint":"Please enter an age between 0 and 100"
+"prequal_select_diagnosis": "Select Diagnosis"
+"prequal_diagnosis_breast_advanced":"Have you been diagnosed with metastatic breast cancer (also known as stage IV breast cancer)?",
+"prequal_diagnosis_prostate_advanced":"Have you been diagnosed with metastatic prostate cancer (also known as prostate cancer that has spread beyond the prostate, including biochemical recurrence)?",
+"prequal_self_age":"How old are you?"
+"prequal_child_age":"How old is your child?"
+"prequal_enter_age":"Enter Age"
+"prequal_age_range":"Please enter an age between 0 and 100"
 
 "cancer": {
   "angiosarcoma": "Angiosarcoma",
@@ -191,13 +191,13 @@
   "liposarcoma": "Liposarcoma",
 },
 
-"prequal_sections_01_blocks_01_nested_01_question_promptTemplate_01"=""
-"prequal_sections_01_blocks_01_nested_01_question_placeholderTemplate_01"="First Name *"
-"prequal_sections_01_blocks_01_nested_01_question_validations_01_hintTemplate_01"="First Name is required"
-"prequal_sections_01_blocks_01_nested_02_question_promptTemplate_01"=""
-"prequal_sections_01_blocks_01_nested_02_question_placeholderTemplate_01"="Last Name *"
-"prequal_sections_01_blocks_01_nested_02_question_validations_01_hintTemplate_01"="Last Name is required"
-"prequal_sections_01_blocks_01_nested_03_question_promptTemplate_01"=""
+"prequal_question1Prompt"=""
+"prequal_firstName"="First Name *"
+"prequal_firstName_required"="First Name is required"
+"prequal_question2Prompt"=""
+"prequal_lastName"="Last Name *"
+"prequal_lastName_required"="Last Name is required"
+"prequal_question3Prompt"=""
 
 #about-you
 "about-you_translatedNames"="About You Survey",
