+++ study-builder/studies/pancan/prequal.conf
@@ -75,11 +75,11 @@
                   "translations": [
                     {
                       "language": "en",
-                      "text": ${i18n.en.prequal_introduction_blocks_01_bodyTemplate_01}
+                      "text": ${i18n.en.prequal_introduction_registration}
                     },
                     {
                       "language": "es",
-                      "text": ${i18n.es.prequal_introduction_blocks_01_bodyTemplate_01}
+                      "text": ${i18n.es.prequal_introduction_registration}
                     }
                   ]
                 }
@@ -99,11 +99,11 @@
                   "translations": [
                     {
                       "language": "en",
-                      "text": ${i18n.en.prequal_introduction_blocks_01_bodyTemplate_02}
+                      "text": ${i18n.en.prequal_introduction_consentForm}
                     },
                     {
                       "language": "es",
-                      "text": ${i18n.es.prequal_introduction_blocks_01_bodyTemplate_02}
+                      "text": ${i18n.es.prequal_introduction_consentForm}
                     }
                   ]
                 }
@@ -123,11 +123,11 @@
                   "translations": [
                     {
                       "language": "en",
-                      "text": ${i18n.en.prequal_introduction_blocks_01_bodyTemplate_03}
+                      "text": ${i18n.en.prequal_introduction_medicalReleaseForm}
                     },
                     {
                       "language": "es",
-                      "text": ${i18n.es.prequal_introduction_blocks_01_bodyTemplate_03}
+                      "text": ${i18n.es.prequal_introduction_medicalReleaseForm}
                     }
                   ]
                 }
@@ -147,11 +147,11 @@
                   "translations": [
                     {
                       "language": "en",
-                      "text": ${i18n.en.prequal_introduction_blocks_01_bodyTemplate_04}
+                      "text": ${i18n.en.prequal_introduction_surveyAnswers}
                     },
                     {
                       "language": "es",
-                      "text": ${i18n.es.prequal_introduction_blocks_01_bodyTemplate_04}
+                      "text": ${i18n.es.prequal_introduction_surveyAnswers}
                     }
                   ]
                 }
@@ -182,11 +182,11 @@
                 "translations": [
                   {
                     "language": "en",
-                    "text": ${i18n.en.prequal_sections_01_blocks_01_title_01}
+                    "text": ${i18n.en.prequal_introduction_titleWho}
                   },
                   {
                     "language": "es",
-                    "text": ${i18n.es.prequal_sections_01_blocks_01_title_01}
+                    "text": ${i18n.es.prequal_introduction_titleWho}
                   }
                 ]
               },
@@ -195,11 +195,11 @@
                 "translations": [
                   {
                     "language": "en",
-                    "text": ${i18n.en.prequal_sections_01_blocks_01_subtitle_01}
+                    "text": ${i18n.en.prequal_Check_that_apply}
                   },
                   {
                     "language": "es",
-                    "text": ${i18n.es.prequal_sections_01_blocks_01_subtitle_01}
+                    "text": ${i18n.es.prequal_Check_that_apply}
                   }
                 ]
               }
@@ -219,11 +219,11 @@
                       "translations": [
                         {
                           "language": "en",
-                          "text": ${i18n.en.prequal_sections_01_blocks_01_nested_03_question_promptTemplate_01}
+                          "text": ${i18n.en.prequal_question3Prompt}
                         },
                         {
                           "language": "es",
-                          "text": ${i18n.es.prequal_sections_01_blocks_01_nested_03_question_promptTemplate_01}
+                          "text": ${i18n.es.prequal_question3Prompt}
                         }
                       ]
                     }
@@ -241,11 +241,11 @@
                           "translations": [
                             {
                               "language": "en",
-                              "text": ${i18n.en.prequal_sections_01_question_01_picklistOptions_01_optionLabelTemplate}
+                              "text": ${i18n.en.prequal_selfDiagnosed}
                             },
                             {
                               "language": "es",
-                              "text": ${i18n.es.prequal_sections_01_question_01_picklistOptions_01_optionLabelTemplate}
+                              "text": ${i18n.es.prequal_selfDiagnosed}
                             }
                           ]
                         }
@@ -266,11 +266,11 @@
                           "translations": [
                             {
                               "language": "en",
-                              "text": ${i18n.en.prequal_sections_01_question_01_picklistOptions_02_optionLabelTemplate}
+                              "text": ${i18n.en.prequal_childDiagnosed}
                             },
                             {
                               "language": "es",
-                              "text": ${i18n.es.prequal_sections_01_question_01_picklistOptions_02_optionLabelTemplate}
+                              "text": ${i18n.es.prequal_childDiagnosed}
                             }
                           ]
                         }
@@ -293,11 +293,11 @@
                           "translations": [
                             {
                               "language": "en",
-                              "text": ${i18n.en.prequal_sections_01_question_01_validations_hintTemplate}
+                              "text": ${i18n.en.prequal_choose_option}
                             },
                             {
                               "language": "es",
-                              "text": ${i18n.es.prequal_sections_01_question_01_validations_hintTemplate}
+                              "text": ${i18n.es.prequal_choose_option}
                             }
                           ]
                         }
