+++ study-builder/studies/pancan/snippets/question-child-age.conf
@@ -12,11 +12,11 @@
         "translations": [
           {
             "language": "en",
-            "text": ${i18n.en.prequal_child_age_prompt}
+            "text": ${i18n.en.prequal_child_age}
           },
           {
             "language": "es",
-            "text": ${i18n.es.prequal_child_age_prompt}
+            "text": ${i18n.es.prequal_child_age}
           }
         ]
       }
@@ -31,11 +31,11 @@
         "translations": [
           {
             "language": "en",
-            "text": ${i18n.en.prequal_age_placeholder}
+            "text": ${i18n.en.prequal_enter_age}
           },
           {
             "language": "es",
-            "text": ${i18n.es.prequal_age_placeholder}
+            "text": ${i18n.es.prequal_enter_age}
           }
         ]
       }
@@ -55,11 +55,11 @@
             "translations": [
               {
                 "language": "en",
-                "text": ${i18n.en.prequal_age_hint}
+                "text": ${i18n.en.prequal_age_range}
               },
               {
                 "language": "es",
-                "text": ${i18n.es.prequal_age_hint}
+                "text": ${i18n.es.prequal_age_range}
               }
             ]
           }
@@ -78,11 +78,11 @@
             "translations": [
               {
                 "language": "en",
-                "text": ${i18n.en.prequal_age_hint}
+                "text": ${i18n.en.prequal_age_range}
               },
               {
                 "language": "es",
-                "text": ${i18n.es.prequal_age_hint}
+                "text": ${i18n.es.prequal_age_range}
               }
             ]
           }
