+++ study-builder/studies/pancan/snippets/question-diagnosed-primary-cancer-breast-advanced.conf
@@ -11,11 +11,11 @@
         "translations": [
           {
             "language": "en",
-            "text": ${i18n.en.prequal_diagnosed_primary_cancer_breast_advanced}
+            "text": ${i18n.en.prequal_diagnosis_breast_advanced}
           },
           {
             "language": "es",
-            "text": ${i18n.es.prequal_diagnosed_primary_cancer_breast_advanced}
+            "text": ${i18n.es.prequal_diagnosis_breast_advanced}
           }
         ]
       }
