+++ study-builder/studies/pancan/snippets/question-diagnosed-primary-cancer-list.conf
@@ -5,13 +5,13 @@
   "childOrientation": "VERTICAL",
   "promptTemplate": {
     "templateType": "HTML",
-    "templateText": "$prequal_diagnosed_primary_cancer_list_prompt"
+    "templateText": "$prequal_diagnosedPrimaryCancerList"
     "variables": [
       {
-        "name": "prequal_diagnosed_primary_cancer_list_prompt",
+        "name": "prequal_diagnosedPrimaryCancerList",
         "translations": [
-          { "language": "en", "text":${i18n.en.prequal_diagnosed_primary_cancer_list_prompt} },
-          { "language": "es", "text": ${i18n.es.prequal_diagnosed_primary_cancer_list_prompt} }
+          { "language": "en", "text":${i18n.en.prequal_diagnosedPrimaryCancerList} },
+          { "language": "es", "text": ${i18n.es.prequal_diagnosedPrimaryCancerList} }
         ]
       }
     ]
@@ -27,11 +27,11 @@
         "translations": [
           {
             "language": "en",
-            "text": ${i18n.en.prequal_diagnosed_primary_cancer_list_question_additionalItemTemplate}
+            "text": ${i18n.en.prequal_diagnose_cancer_list_additionalTmpl}
           },
           {
             "language": "es",
-            "text": ${i18n.es.prequal_diagnosed_primary_cancer_list_question_additionalItemTemplate}
+            "text": ${i18n.es.prequal_diagnose_cancer_list_additionalTmpl}
           }
         ]
       }
@@ -47,11 +47,11 @@
         "translations": [
           {
             "language": "en",
-            "text": ${i18n.en.prequal_diagnosed_primary_cancer_list_add_buttonTemplate}
+            "text": ${i18n.en.prequal_diagnosed_cancer_list_additional_diagnosis}
           },
           {
             "language": "es",
-            "text": ${i18n.es.prequal_diagnosed_primary_cancer_list_add_buttonTemplate}
+            "text": ${i18n.es.prequal_diagnosed_cancer_list_additional_diagnosis}
           }
         ]
       }
