+++ study-builder/studies/pancan/snippets/question-diagnosed-primary-cancer.conf
@@ -3,13 +3,13 @@
   "stableId": "DIAGNOSED_PRIMARY_CANCER",
   "promptTemplate": {
     "templateType": "HTML",
-    "templateText": "$prequal_diagnosed_primary_cancer_prompt"
+    "templateText": "$prequal_select_diagnosis"
     "variables": [
       {
-        "name": "prequal_diagnosed_primary_cancer_prompt",
+        "name": "prequal_select_diagnosis",
         "translations": [
-          { "language": "en", "text":${i18n.en.prequal_diagnosed_primary_cancer_prompt} },
-          { "language": "es", "text": ${i18n.es.prequal_diagnosed_primary_cancer_prompt} }
+          { "language": "en", "text":${i18n.en.prequal_select_diagnosis} },
+          { "language": "es", "text": ${i18n.es.prequal_select_diagnosis} }
         ]
       }
     ]
