+++ study-builder/studies/pancan/study-workflows.conf
@@ -27,7 +27,7 @@
         },
         {
           "type": "STUDY_REDIRECT",
-          "studyGuid": ${mbcStudyGuid},
+          "studyGuid": "cmi-mbc",
           "redirectUrl": ${mbcRedirectUrl},
           ## TODO update PEX to add && user.studies["cmi-pancan"].forms["PREQUAL"].questions["DIAGNOSED_PRIMARY_CANCER_LIST"].answers.numSelectedOptions() == 1
           ## and user.profile.language() == "en"
@@ -42,7 +42,7 @@
 ## MBC spanish site redirect
 ##        {
 ##          "type": "STUDY_REDIRECT",
-##          "studyGuid": ${mbcStudyGuid},
+##          "studyGuid": "cmi-mbc",
 ##          "redirectUrl": ${mbcRedirectUrlES},
 ##          "expression": """
 ##              user.studies["cmi-pancan"].forms["PREQUAL"].isStatus("COMPLETE")
@@ -55,7 +55,7 @@
 
         {
           "type": "STUDY_REDIRECT",
-          "studyGuid": ${mpcStudyGuid},
+          "studyGuid": "cmi-mpc",
           "redirectUrl": ${mpcRedirectUrl},
           ## TODO update PEX
           "expression": """
@@ -66,7 +66,7 @@
         },
         {
           "type": "STUDY_REDIRECT",
-          "studyGuid": ${brainStudyGuid},
+          "studyGuid": "cmi-brain",
           "redirectUrl": ${brainRedirectUrl},
           ## TODO update PEX
           "expression": """
@@ -76,7 +76,7 @@
         },
         {
           "type": "STUDY_REDIRECT",
-          "studyGuid": ${angioStudyGuid},
+          "studyGuid": "ANGIO",
           "redirectUrl": ${angioRedirectUrl},
           ## TODO update PEX
           "expression": """
@@ -86,7 +86,7 @@
         },
         {
           "type": "STUDY_REDIRECT",
-          "studyGuid": ${osteoStudyGuid},
+          "studyGuid": "CMI-OSTEO",
           "redirectUrl": ${osteoRedirectUrl},
           ## TODO update PEX
           "expression": """
